[#11930] Extracting to a separate file
